+++ NEWS.md
@@ -1,4 +1,11 @@
-# mlr3misc 0.13.0-9000
+# mlr3misc (development version)
+
+# mlr3misc 0.15.0
+
+* Feat: Added `strip_screfs` S3 generic, which removes source references from objects
+* The RNG state is now copied to the callr session when using `encapsulate()`.
+
+# mlr3misc 0.14.0
 
 * Added argument `.compile` to function `crate()` because R disables byte-code
 compilation of functions when changing their enclosing environment